requirements_pro.txt
# Professional Production Requirements for RAG System

# Core RAG Dependencies
langchain==0.1.17
langchain-community==0.0.37
langchain-openai==0.1.7
langchain-chroma==0.1.1

# Document Processing
pypdf2==3.0.1
pymupdf==1.24.2  # fitz for better PDF extraction
pytesseract==0.3.10  # OCR support
pillow==10.3.0

# Vector Database
chromadb==0.4.24
faiss-cpu==1.8.0

# Embeddings and LLM
openai==1.30.1
sentence-transformers==2.7.0
transformers==4.40.1
torch==2.3.0

# Multilingual Support
polyglot==16.7.4
pycld2==0.41  # Language detection
googletrans==4.0.0rc1  # Translation if needed

# Text Processing
nltk==3.8.1
spacy==3.7.4
indic-nlp-library==0.81  # Bengali text processing
bnlp==2.0.2  # Bengali NLP toolkit

# API Framework - Enhanced
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0  # Production WSGI server
pydantic==2.7.1

# Security & Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.7

# Caching & Storage
redis==5.0.6
hiredis==2.3.2  # Faster Redis client
psycopg2-binary==2.9.9  # PostgreSQL adapter
sqlalchemy==2.0.30
alembic==1.13.1  # Database migrations

# Monitoring & Metrics
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.3.1
psutil==5.9.8
py-cpuinfo==9.0.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Evaluation
rouge-score==0.1.2
bert-score==0.3.13
sacrebleu==2.4.2

# HTTP & Networking
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.3

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4
regex==2024.4.28
click==8.1.7

# Logging & Configuration
structlog==24.1.0
pyyaml==6.0.1
toml==0.10.2

# Development & Testing (optional - install with pip install -r requirements_dev.txt)
# pytest==8.2.1
# pytest-asyncio==0.23.7
# pytest-cov==5.0.0
# black==24.4.2
# flake8==7.0.0
# mypy==1.10.0
# pre-commit==3.7.1

# Performance & Optimization
cython==3.0.10
numba==0.59.1

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Validation & Serialization
marshmallow==3.21.3
cerberus==1.3.5

# Background Tasks
celery==5.3.7
kombu==5.3.7

# Rate Limiting
slowapi==0.1.9

# Environment & Configuration Management
pydantic-settings==2.3.2
environs==11.0.0
